db moved inside the server folder, default.png constant extracted
