+++ server/forms/signin.go
@@ -2,7 +2,7 @@ package forms
 
 import (
 	"github.com/badoux/checkmail"
-	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/db"
+	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/db"
 	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/formats"
 )
