+++ server/forms/signup.go
@@ -3,7 +3,7 @@ package forms
 import (
 	"errors"
 	"github.com/badoux/checkmail"
-	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/db"
+	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/db"
 	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/formats"
 	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/settings"
 	"github.com/lib/pq"
