+++ server/routes/user_get_test.go
@@ -3,7 +3,7 @@ package routes
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/db"
+	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/db"
 	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/formats"
 	. "github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/test-utils"
 	"io/ioutil"
