+++ server/routes/user_put.go
@@ -3,7 +3,7 @@ package routes
 import (
 	"encoding/json"
 	"github.com/disintegration/imaging"
-	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/db"
+	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/db"
 	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/formats"
 	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/forms"
 	. "github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/handlers"
