+++ server/routes/validators.go
@@ -2,7 +2,7 @@ package routes
 
 import (
 	"errors"
-	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/db"
+	"github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/db"
 	"net/url"
 	"strconv"
 	"strings"
