+++ server/test-utils/mock_data.go
@@ -2,7 +2,7 @@ package test_utils
 
 import (
 	"database/sql"
-	. "github.com/go-park-mail-ru/2019_1_three_in_a_boat/db"
+	. "github.com/go-park-mail-ru/2019_1_three_in_a_boat/server/db"
 	"time"
 )
